leerie: Extract duplicated _stub_self_cmd into tests/stub_helpers.py
test_chain_launcher_id_dispatch.py, test_group_launcher.py, and
test_group_launcher_verbs.py each defined an identical bash-stub
argv-recorder for intercepting LEERIE_SELF_CMD dispatch in chain/
group launcher tests.
