+++ tests/test_group_launcher_verbs.py
@@ -20,8 +20,9 @@ from __future__ import annotations
 import json
 import os
 import subprocess
-import textwrap
 from pathlib import Path
+
+from tests.stub_helpers import _stub_self_cmd
 
 REPO_ROOT = Path(__file__).resolve().parent.parent
 LAUNCHER = REPO_ROOT / "leerie"
@@ -43,22 +44,6 @@ def _write_run(state_dir: Path, run_id: str, fields: dict) -> None:
     rd = state_dir / "runs" / run_id
     rd.mkdir(parents=True)
     (rd / "run.json").write_text(json.dumps(fields))
-
-
-def _stub_self_cmd(tmp_path: Path) -> tuple[Path, Path]:
-    """Stub binary that records its full argv per invocation.
-
-    Returns ``(stub_path, log_path)``.
-    """
-    log = tmp_path / "stub-self.log"
-    stub = tmp_path / "self-stub"
-    stub.write_text(textwrap.dedent(f"""\
-        #!/usr/bin/env bash
-        echo "$@" >> "{log}"
-        exit 0
-        """))
-    stub.chmod(0o755)
-    return stub, log
 
 
 def _two_member_fixture(tmp_path: Path) -> tuple[Path, Path]:
